faed_management_tool/requirements.txt
Django
djangorestframework
requests
simplekml
polycircles